fix loading bug
Some pixelMaps were mysteriously empty. Maybe that's because drawing
an image to the canvas and then immediately reading image data from
the canvas isn't supported, we need to give control back to the
browser in order for it to complete the drawing action? Trying this
required a big refactoring, and it did fix the problem, but loading is
super slow now :(