prevent flickering
When computing the collisions, we don't know whether there are any
collisions yet, so I default to "yes" so players can't abuse this to
skip the level. This led to a flickering effect where each change
caused the collision state to quickly switch from "yes" to "no". Now
we still default to "yes", but we display the last known state instead
of displaying "yes" when unknown, thereby eliminating the flickering.